+++ ui/src/notebooks/components/Notebook.tsx
@@ -4,7 +4,7 @@ import {Page} from '@influxdata/clockface'
 import {NotebookProvider} from 'src/notebooks/context/notebook'
 import Header from 'src/notebooks/components/Header'
 import PipeList from 'src/notebooks/components/PipeList'
-import AddButtons from 'src/notebooks/components/AddButtons'
+import NotebookPanel from 'src/notebooks/components/panel/NotebookPanel'
 
 // NOTE: uncommon, but using this to scope the project
 // within the page and not bleed it's dependancies outside
@@ -15,16 +15,14 @@ const NotebookPage: FC = () => {
   return (
     <NotebookProvider>
       <Page titleTag="Notebook">
-        <Page.Header fullWidth={false}>
-          <Header />
-        </Page.Header>
-        <Page.Contents fullWidth={false} scrollable={true}>
+        <Header />
+        <Page.Contents fullWidth={true} scrollable={true}>
           <PipeList />
-          <AddButtons />
         </Page.Contents>
       </Page>
     </NotebookProvider>
   )
 }
 
+export {NotebookPanel}
 export default NotebookPage